sesion8/microservicios/alert/requirements.txt
flask
httpx
python-dotenv
asyncio